rearranged files: created a distinct sub-module 'fmi' which contains the fmi dataclasses
